Compile fixes...
<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ client/src/ui/workbench/workbench_access_controller.cpp
@@ -81,7 +81,7 @@ Qn::GlobalPermissions QnWorkbenchAccessController::globalPermissions(const QnUse
     if(!user)
         return result;
 
-    result = static_cast<Qn::GlobalPermissions>(user->getPermissions());
+    result = static_cast<Qn::GlobalPermissions>(static_cast<int>(user->getPermissions()));
 
     if(user->isAdmin())
         result |= Qn::GlobalOwnerPermissions;
